Add sharing options

+++ java-tools/OsmAndServer/src/main/java/net/osmand/server/ws/TranslationMessage.java
@@ -7,8 +7,8 @@ public class TranslationMessage {
 	public static final String TYPE_MSG_JOIN = "join";
 	
 	public static final String SENDER_SYSTEM = "System";
-	public static final long SENDER_SYSTEM_ID = 0;
-	public static final long SENDER_ANONYMOUS_ID = -1;
+	public static final int SENDER_SYSTEM_ID = 0;
+	public static final int SENDER_ANONYMOUS_ID = -1;
 
 	public String sender;
 	public long sendDeviceId;
